+++ CHANGELOG.md
@@ -20,9 +20,34 @@
   - [v0.2.0](services/modelexperiments/CHANGELOG.md#v020)
     - **New**: STACKIT Model Experiments module wait handler added.
 - `rabbitmq`:
-  - [v1.1.0](/services/rabbitmq/CHANGELOG.md#v110)
+  - [v1.1.0](services/rabbitmq/CHANGELOG.md#v110)
    - `v2api`:
     - **Feature**: Added wait handlers
+- `ske`:
+  - [v1.19.0](services/ske/CHANGELOG.md#v1190)
+    - Package `v1api`:
+      - **Improvement:** Add validation to Name field of model struct `Cluster`
+    - Package `v2api`:
+      - **Improvement:** Add validation to Name field of model struct `Cluster`
+      - **Feature:** New model struct `Audit`
+      - **Feature:** Add Audit (type Audit) field to model struct `CreateOrUpdateClusterPayload`
+    - Deprecated SDK layer in root of the module:
+      - **Improvement:** Add validation to Name field of model struct `Cluster`
+      - **Feature:** New model struct `Audit`
+      - **Feature:** Add Audit (type Audit) field to model struct `CreateOrUpdateClusterPayload`
+- `sqlserverflex`: 
+  - [v1.14.0](services/sqlserverflex/CHANGELOG.md#v1140)
+    - `v3beta2api`: **New:** New package which can be used for communication with the sqlserverflex v3beta2 API
+    - `v1api`: **Deprecated:** `v1api` is deprecated, use instead `v2api`
+    - `v3alpha1api`: **Deprecated:** `v3alpha1api` is deprecated, use instead `v3beta2api`
+    - `v3beta1api`: **Deprecated:** `v3beta1api` is deprecated, use instead `v3beta2api`
+- `telemetrylink`:
+  - [v0.3.0](services/telemetrylink/CHANGELOG.md#v030)
+    - **New:** v1api API version for STACKIT Telemetry Link
+- `telemetryrouter`:
+  - [v0.4.0](services/telemetryrouter/CHANGELOG.md#v040)
+    - **New**: v1api API version for STACKIT Telemetry Router
+
 
 ## Release (2026-06-18)
 - `core`: