+++ services/telemetrylink/CHANGELOG.md
@@ -1,3 +1,6 @@
+## v0.3.0
+- **New:** v1api API version for STACKIT Telemetry Link
+
 ## v0.2.0
 - **Feature:** Introduce enums for various attributes